Switch to mySQL

+++ shopyP/models.py
@@ -18,12 +18,11 @@ def load_user(user_id):
 
 # Parent class for User + Admin
 class Person():
-    id = db.Column(db.Integer, primary_key=True)
+    id = db.Column(db.BigInteger, primary_key=True)
     username = db.Column(db.String(20), unique=True, nullable=False)
     email = db.Column(db.String(120), unique=True, nullable=False)
     image_file = db.Column(db.String(20), nullable=False, default='default.jpg')
     password = db.Column(db.String(60), nullable=False)
-
 
 
     def __repr__(self):
